JSON should be displayed now

+++ Backend/timperio.crm/src/main/java/oopsies/timperio/crm/Repository/CustomerRepository.java
@@ -5,6 +5,9 @@ import org.springframework.stereotype.Repository;
 import oopsies.timperio.crm.Customer; 
 
 @Repository
-public interface CustomerRepository extends JpaRepository<Customer, Long> {
+public interface CustomerRepository extends JpaRepository<Customer, Integer> { // Change Long to Integer
     // Custom query methods (if needed) can be added here
 }
+
+
+
